Tests/LibWeb/Text/expected/css/CSSStyleRule-pseudoElement-selectorText.txt

::-webkit-scrollbar-thumb
::-webkit-scrollbar-thumb:hover
